fix: error if proper file exist

+++ src/fs/rename.js
@@ -12,9 +12,24 @@ const rename = async () => {
     try {
         const pathToWrongFile = path.join(__dirname, 'files', WRONG_FILENAME);
         const pathToRenamedFile = path.join(__dirname, 'files', PROPER_FILENAME);
-        await fs.rename(pathToWrongFile, pathToRenamedFile)
+        const isExist = await chechIfFileExist(pathToRenamedFile);
+        if (!isExist) {
+            await fs.rename(pathToWrongFile, pathToRenamedFile)
+        } else {
+            throw new FSError();
+        }
+       
     } catch {
         throw new FSError();
+    }
+};
+
+const chechIfFileExist = async (path) => {
+    try {
+        await fs.access(path);
+        return true;
+    } catch {
+        return false;
     }
 };
